+++ CHANGELOG.md
@@ -776,3 +776,4 @@
 - [2026-04-29 18:55] feat(i18n): implement responsive i18n view component
 - [2026-04-29 19:21] feat(storage): integrate custom telemetry logger in storage
 - [2026-04-29 19:37] feat(forms): integrate custom telemetry logger in forms
+- [2026-04-29 20:24] feat(core): add async core stream provider